+++ pybamm/models/submodels/thermal/pouch_cell/pouch_cell_2D_current_collectors.py
@@ -60,34 +60,61 @@ class CurrentCollector2D(BaseThermal):
         # Account for surface area to volume ratio of pouch cell in cooling
         # coefficient. Note: the factor 1/delta^2 comes from the choice of
         # non-dimensionalisation
-        A = self.param.l_y * self.param.l_z
-        V = self.param.l * self.param.l_y * self.param.l_z
-        cooling_coeff = -2 * self.param.h * A / V / (self.param.delta ** 2)
+        yz_surface_area = self.param.l_y * self.param.l_z
+        cell_volume = self.param.l * self.param.l_y * self.param.l_z
+        yz_surface_cooling_coefficient = (
+            -(self.param.h_cn + self.param.h_cp)
+            * yz_surface_area
+            / cell_volume
+            / (self.param.delta ** 2)
+        )
 
-        # Add boundary source term which accounts for surface cooling around
-        # the edge of the domain in the weak formulation.
-        # TODO: update to allow different cooling conditions at the tabs
+        edge_cooling_coefficient = self.param.h_edge / self.param.delta
+
+        # Governing equations contain:
+        #   - source term for y-z surface cooling
+        #   - boundary source term of edge cooling
+        # Boundary conditions contain:
+        #   - Neumann condition for tab cooling
         self.rhs = {
             T_av: (
                 pybamm.laplacian(T_av)
                 + self.param.B * pybamm.source(Q_av, T_av)
-                + cooling_coeff * pybamm.source(T_av - T_amb, T_av)
-                - (self.param.h / self.param.delta)
+                + yz_surface_cooling_coefficient * pybamm.source(T_av - T_amb, T_av)
+                - edge_cooling_coefficient
                 * pybamm.source(T_av - T_amb, T_av, boundary=True)
             )
             / (self.param.C_th * self.param.rho)
         }
 
+        # TODO: Make h_edge a function of position to have bottom/top/side cooled cells.
+
     def set_boundary_conditions(self, variables):
         T_av = variables["X-averaged cell temperature"]
-        # Dummy no flux boundary conditions since cooling at the the tabs is
-        # accounted for in the boundary source term in the weak form of the
-        # governing equation
-        # TODO: update to allow different cooling conditions at the tabs
+        T_amb = variables["Ambient temperature"]
+
+        # Subtract the edge cooling from the tab portion so as to not double count
+        # Note: tab cooling is also only applied on the current collector hence
+        # the (l_cn / l) and (l_cp / l) prefactors.
+        # We also still have edge cooling on the region: x in (0, 1)
+        h_tab_n_corrected = (
+            (self.param.l_cn / self.param.l)
+            * (self.param.h_tab_n - self.param.h_edge)
+            / self.param.delta
+        )
+        h_tab_p_corrected = (
+            (self.param.l_cp / self.param.l)
+            * (self.param.h_tab_p - self.param.h_edge)
+            / self.param.delta
+        )
+
+        T_av_n = pybamm.BoundaryValue(T_av, "negative tab")
+        T_av_p = pybamm.BoundaryValue(T_av, "positive tab")
+
         self.boundary_conditions = {
             T_av: {
-                "negative tab": (pybamm.Scalar(0), "Neumann"),
-                "positive tab": (pybamm.Scalar(0), "Neumann"),
+                "negative tab": (-h_tab_n_corrected * (T_av_n - T_amb), "Neumann"),
+                "positive tab": (-h_tab_p_corrected * (T_av_p - T_amb), "Neumann"),
             }
         }